Inventario_Ventas-Backend/requirements.txt
asgiref==3.7.2
Django==4.2.6
django-cors-headers==4.4.0
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3

#user: admin
#password: admin